+++ CHANGELOG.md
@@ -6,8 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## Unreleased
 
+--- 
+## [2.0.3] - 2025-09-22
+## Changed
+- **Breaking Change**: Determinism between C_reference and Python 
+    - `fpsr_algorithms.py` and `fpsr_algorithms_reference.c` now 
+        - produce the same random value outputs for the same seed and parameters.
+        - values jump on the same frames
+    - **Breaks compatibility** from `2.02`
+        - In previous versions, values generated within the same languages are deterministic, repeatable within their own language implementation, but were not deterministic across languages.
+        - This was due to different ways C and Python were handling large integers and modulo operations for negative numbers.
+    - FPS-R in other languages are still deterministic with themselves, but they do not guarantee consistency with C (reference source of truth) and other applications.
+- Wrapper versions are placed back to `./resources/code/_WIP` to be re-worked since they are not fully working.
 
-## [2.02] - 2025-09-21
+---
+## [2.0.2] - 2025-09-21
 ### Added
 - an English podcast style introduction to FPS-R
 
@@ -19,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [2.0.1] - 2025-09-14
 ### Added 
 - A wrapper version of each FPS-R function (SM, TM, QS):
+    console.log("🚀 ~ FPS:", FPS)
     - enables rich analytic information on function output:
         - `has_changed` checking with the value output of previous frame, returning `1` or `True` if value has changed or "jumped"
         - `hold_progress` Normalized progress (0.0 to 1.0) through the current hold duration.